+++ sbfrm_gui.py
@@ -16,59 +16,64 @@ def make_window():
                            default=last_up_col, key="-UPDATE_COLLECTION-"),
                   sg.Radio('Update System', "-RADIO-",
                            default=last_up_sys, key="-UPDATE_SYSTEM-")]]
+
     last_src = sg.user_settings_get_entry('-last source-', '')
-    select_src = [sg.Text("     Source", size=(15, 1)),
+    select_src = [sg.Text("     Source", size=(13, 1)),
                   sg.Input(last_src, key='-SOURCE-', enable_events=True),
                   sg.FolderBrowse(initial_folder=last_src)]
+
     last_dest = sg.user_settings_get_entry('-last dest-', '')
-    select_dest = [sg.Text("     Destination", size=(15, 1)),
+    select_dest = [sg.Text("     Destination", size=(13, 1)),
                    sg.Input(last_dest, key='-DEST-'),
                    sg.FolderBrowse(initial_folder=last_dest)]
-    select_boxart = [sg.Text("     Boxart", size=(15, 1)),
+
+    select_boxart = [sg.Text("     Boxart", size=(13, 1)),
                      sg.Input(sg.user_settings_get_entry('-last boxarts-', ''),
-                     key='-BOXART-', enable_events=True),
-                     sg.FolderBrowse()]
-    select_img = [sg.Text("     Image", size=(15, 1)),
+                     key='-BOXART-', size=(57, 1))]
+
+    select_img = [sg.Text("     Image", size=(13, 1)),
                   sg.Input(sg.user_settings_get_entry('-last images-', ''),
-                  key='-IMAGE-', enable_events=True),
-                  sg.FolderBrowse()]
-    select_thumb = [sg.Text("     Thumbnail", size=(15, 1)),
+                  key='-IMAGE-', size=(57, 1))]
+
+    select_thumb = [sg.Text("     Thumbnail", size=(13, 1)),
                     sg.Input(sg.user_settings_get_entry('-last thumbs-', ''),
-                    key='-THUMB-', enable_events=True),
-                    sg.FolderBrowse()]
-    select_marc = [sg.Text("     Marquee", size=(15, 1)),
+                    key='-THUMB-', size=(57, 1))]
+
+    select_marc = [sg.Text("     Marquee", size=(13, 1)),
                    sg.Input(sg.user_settings_get_entry('-last marquees-', ''),
-                            key='-MARQUEE-',
-                            enable_events=True),
-                   sg.FolderBrowse()]
-    select_vid = [sg.Text("     Video", size=(15, 1)),
+                   key='-MARQUEE-', size=(57, 1))]
+
+    select_vid = [sg.Text("     Video", size=(13, 1)),
                   sg.Input(sg.user_settings_get_entry('-last videos-', ''),
-                  key='-VIDEO-', enable_events=True),
-                  sg.FolderBrowse()]
-    subsystems = [sg.Text("     Subsystems", size=(15, 1)),
+                  key='-VIDEO-', size=(57, 1))]
+
+    subsystems = [sg.Text("     Subsystems", size=(13, 1)),
                   sg.Multiline(sg.user_settings_get_entry('-last subsystems-', ''),
-                  key='-SUBSYSLIST-', size=(42, 6))]
+                  key='-SUBSYSLIST-', size=(55, 6))]
     layout = [
         [sg.Text("Choose Operation.", font=("Helvetica", 12))],
         operation,
         [sg.Text("Set Source and Destination Paths.", font=("Helvetica", 12))],
         select_src,
         select_dest,
-        [sg.Text("Set the System Media Folder Source Names.",
-                 font=("Helvetica", 12))],
+        [sg.Text("Source Media Folder Names.", font=("Helvetica", 12)),
+         sg.Text("(Comma separated, no spaces after comma)",
+                 font=("Helvetica", 8),
+                 text_color='dark gray')],
         select_boxart,
         select_img,
         select_thumb,
         select_marc,
         select_vid,
-        [sg.Text("Set the SubSystems Names.", font=("Helvetica", 12)),
-         sg.Text("(One per line, no spaces)",
-                 font=("Helvetica", 8), text_color='dark gray')],
+        [sg.Text("SubSystems Names.", font=("Helvetica", 12)),
+         sg.Text("(Comma separated, no spaces after comma, multiple lines)",
+                 font=("Helvetica", 8),
+                 text_color='dark gray')],
         subsystems,
         [sg.Button('Ok'), sg.Button('Exit')],
     ]
 
-    return sg.Window('SBFRM V0.4.2', layout, enable_close_attempted_event=True,
+    return sg.Window('SBFRM V0.4.3', layout, enable_close_attempted_event=True,
                      finalize=True,
                      location=sg.user_settings_get_entry('-location-',
                                                          (None, None)))
@@ -78,11 +83,11 @@ def make_progress_window(location):
     """Make Description"""
     layout = [
         [sg.Output(size=(72, 6), key='-CONSOLE-')],
-        [sg.Text("Collection Progress:", size=(15, 1)), sg.ProgressBar(
-            max_value=10, orientation='h', size=(38, 10),
+        [sg.Text("Collection Progress:", size=(16, 1)), sg.ProgressBar(
+            max_value=10, orientation='h', size=(37, 10),
             key='-PROGRESS_SYSTEMS-')],
-        [sg.Text("System Progress:", size=(15, 1)), sg.ProgressBar(
-            max_value=10, orientation='h', size=(38, 10),
+        [sg.Text("System Progress:", size=(16, 1)), sg.ProgressBar(
+            max_value=10, orientation='h', size=(37, 10),
             key='-PROGRESS_GAMES-')],
         [sg.Button('Cancel')]
     ]
@@ -97,7 +102,22 @@ def update_collection(collection, values, gui):
              "marquee": values['-THUMB-'],
              "thumbnail": values['-MARQUEE-'],
              "video": values['-VIDEO-']}
+    for key, value in mdirs.items():
+        while ', ' in value:
+            value = value.replace(', ', ',')
+        while value[-1] in [',', ' ']:
+            value = value[:-1]
+        mdirs[key] = value.split(',')
+
     subsyslist = values['-SUBSYSLIST-'].splitlines()
+    for subsysid, value in enumerate(subsyslist):
+        while ', ' in value:
+            value = value.replace(', ', ',')
+        while value[-1] in [',', ' ']:
+            value = value[:-1]
+        subsyslist[subsysid] = value.split(',')
+
+    subsyslist = [s for ls in subsyslist for s in ls]
     print(60*'='+'\n', 'Update Collection:')
 
     sys_paths = collection.list_systems(values['-SOURCE-'])
@@ -118,13 +138,20 @@ def update_collection(collection, values, gui):
 def update_sys_from(collection, values, gui):
     """Make Description"""
     print('Update System:', values['-DEST-'])
-    media_dirs = {"boxart": values['-BOXART-'],
-                  "image": values['-IMAGE-'],
-                  "marquee": values['-THUMB-'],
-                  "thumbnail": values['-MARQUEE-'],
-                  "video": values['-VIDEO-']}
+    mdirs = {"boxart": values['-BOXART-'],
+             "image": values['-IMAGE-'],
+             "marquee": values['-THUMB-'],
+             "thumbnail": values['-MARQUEE-'],
+             "video": values['-VIDEO-']}
+    for key, value in mdirs.items():
+        while ', ' in value:
+            value = value.replace(', ', ',')
+        while value[-1] in [',', ' ']:
+            value = value[:-1]
+        mdirs[key] = value.split(',')
+
     collection.update_sys_from(values['-DEST-'], values['-SOURCE-'],
-                               media_dirs, gui=gui)
+                               mdirs, gui=gui)
 
     gui.write_event_value('-UPDATE_SYSTEM_END-', '')
 
@@ -175,10 +202,6 @@ def main():
             sg.user_settings_set_entry('-last source-', values['-SOURCE-'])
             sg.user_settings_set_entry('-last dest-', values['-DEST-'])
 
-        if event in ['-BOXART-', '-IMAGE-', '-THUMB-', '-MARQUEE-', '-VIDEO-']:
-            values[event] = basename(values[event])
-            window[event].update(values[event])
-
         if event == 'Ok' and values['-UPDATE_COLLECTION-'] is True:
             xpos, ypos = window1.current_location()
             xpos += 30
@@ -202,7 +225,9 @@ def main():
             thread.start()
 
         if event == 'Cancel':
-            print("Halting the Process. Please wait...")
+            print("\n\n=============================================",
+                  "\nHalting the Process. Please wait...",
+                  "\n=============================================")
             col.stop_copy = True
 
         if event == '-PROGRESS_GAMES-':